Core: Updated to 11.2.5.63704
Port From (https://github.com/TrinityCore/TrinityCore/commit/53068a94e88991ae7196fb247df4b7ca2e47e554)

+++ Source/Game/Calendar/CalendarManager.cs
@@ -539,7 +539,7 @@ namespace Game
             CalendarInviteRemoved packet = new();
             packet.ClearPending = true; // FIXME
             packet.EventID = calendarEvent.EventId;
-            packet.Flags = flags;
+            packet.Flags = (ushort)flags;
             packet.InviteGuid = invite.InviteeGuid;
 
             SendPacketToAllEventRelatives(packet, calendarEvent);
