+++ Source/Game/Networking/Packets/LFGPackets.cs
@@ -167,7 +167,7 @@ namespace Game.Networking.Packets
             Ticket.Write(_worldPacket);
 
             _worldPacket.WriteUInt8(SubType);
-            _worldPacket.WriteUInt8(Reason);
+            _worldPacket.WriteUInt32(Reason);
             _worldPacket.WriteInt32(Slots.Count);
             _worldPacket.WriteUInt8(RequestedRoles);
             _worldPacket.WriteInt32(SuspendedPlayers.Count);
@@ -190,7 +190,7 @@ namespace Game.Networking.Packets
 
         public RideTicket Ticket = new();
         public byte SubType;
-        public byte Reason;
+        public uint Reason;
         public List<uint> Slots = new();
         public byte RequestedRoles;
         public List<ObjectGuid> SuspendedPlayers = new();
